refactor: post.goにPostに関する処理を移動
